Split entrypoints for node and browser
