+++ src/config.ts
@@ -5,15 +5,15 @@ export interface Config {
   townCacheSize: number
 
   /** 都道府県＋市区町村のデータをファイルシステムにキャッシュするかどうか */
-  preloadTownCache: boolean
+  preloadCache?: boolean
 
   /** 都道府県＋市区町村のファイルシステムキャッシュの有効期限(単位: ミリ秒) */
-  preloadedTownCacheExpiresIn: number
+  preloadedCacheExpiresIn: number
 }
 
 export const currentConfig: Config = {
   japaneseAddressesApi: 'https://geolonia.github.io/japanese-addresses/api/ja',
   townCacheSize: 1_000,
-  preloadTownCache: false,
-  preloadedTownCacheExpiresIn: 24 * 60 * 60 * 1000,
+  preloadCache: false,
+  preloadedCacheExpiresIn: 24 * 60 * 60 * 1000,
 }
