+++ src/lib/fetch/browser.ts
@@ -1,0 +1,7 @@
+import { currentConfig } from '../../config'
+import unfetch from 'isomorphic-unfetch'
+import { DataFetcher } from '../../normalize'
+
+export const fetchData: DataFetcher = (url: string) => {
+  return unfetch(`${currentConfig.japaneseAddressesApi}${encodeURI(url)}`)
+}
