+++ src/main-node.ts
@@ -1,0 +1,5 @@
+import { createNormalizer } from './normalize'
+import { fetchLocalData, preloadJapaneseAddresses } from './lib/fetch/node'
+
+export const normalize = createNormalizer(fetchLocalData)
+export const preload = preloadJapaneseAddresses
